Streamer throughput optimizations (supersedes #9632) (#19156)
* The Windows Storage Drive in AZ::IO::Streamer now more aggressively queues async read requests.

This reduced the gap between reads from an initial batch when no other reads are running. This yielded around 2% performance gain. (5-10Mib/s on a SATA SSD.)

Signed-off-by: AMZN-koppersr <82230785+AMZN-koppersr@users.noreply.github.com>

* Resolve AZ::IO::Streamer's RequestPath upon construction.

Resolving relative paths and aliases to absolute paths is now done when a RequestPath is constructed instead of using lazy initialization. Practically this means that AZ::IO::Streamer rarely has to resolve paths and the code calling into Streamer will instead resolve the path. Overall this shaved off about 5 to 10 microseconds and reduced the gap between the drive read speed and the overall read speed with Streamer by 1-2Mib/s (on a SATA SSD).

Signed-off-by: AMZN-koppersr <82230785+AMZN-koppersr@users.noreply.github.com>

* Completion callbacks for external requests to AZ::IO::Streamer are now ran in a TaskGraph.

When finalizing requests all the external requests are being queued on a task graph for later execution. This has the added benefit that long running callbacks no longer negatively impact AZ::IO::Streamer. The downside is that cheap callbacks, those that would be faster than creating and queuing a task, are now going to be more expensive. On average though this is still a net gain.

Signed-off-by: AMZN-koppersr <82230785+AMZN-koppersr@users.noreply.github.com>

* A default callback is no longer set on AZ::IO::Streamer requests.

This prevents tasks being created for external requests that don't have a callback and instead the request and any dependencies can immediately be processed. This also slightly improved performance for internal requests.

Signed-off-by: AMZN-koppersr <82230785+AMZN-koppersr@users.noreply.github.com>

* AZ::IO::Streamer now issues read requests through the task system.

Issuing read request with the OS was taking up the majority of the read time (when no files needed to be opened). By moving the actual call to the OS to the task system the scheduler can work on the next request in the meantime and allows issuing read requests to the OS to overlap. In the test scenario this marginally improved the read speed of the drive, but closed gap between the overall throughput of Streamer and the drive to less than 1%.
While the test scenario didn't provide a significant improvement, speeds running in developer mode (--cl_streamerDevMode=true) have more than quadrupled. It should be noted that in test case the same few files get repeatedly read from the disk, which means in dev mode the Windows file cache is mostly used, but this doesn't translate to similar performance in real world scenarios as files are not frequently loaded within a time frame that the file cache would be beneficial (especially with the Asset Manager's aggressive caching of assets). It also doesn't perfectly demonstrates the ability of NVMe drives as the read request calls to the OS are much longer than normal due to copying from the cache. It does however show a significant improvement in handling fast reads, demonstrating a better chance of Streamer keeping up with faster driver.

Signed-off-by: AMZN-koppersr <82230785+AMZN-koppersr@users.noreply.github.com>

* Fixed the AZ::IO::Streamer unit tests that were broken due to the recent optimizations.

Signed-off-by: AMZN-koppersr <82230785+AMZN-koppersr@users.noreply.github.com>

* Addressed PR feedback for AZ::IO::Streamer throughput optimizations.

Changes mostly include code cleanup that makes the code easier to read and adhere to code standards. A few changes are going to have a very minor positive impact on performance as well.

Signed-off-by: AMZN-koppersr <82230785+AMZN-koppersr@users.noreply.github.com>

* Fixed a compile error on Android due to std::atomic not conforming to the standard. (value_type is missing.)

Signed-off-by: AMZN-koppersr <82230785+AMZN-koppersr@users.noreply.github.com>

* Added a missing Task Executor to a unit test to prevent it from crashing.

Signed-off-by: AMZN-koppersr <82230785+AMZN-koppersr@users.noreply.github.com>

* Fixed more tests that needed a TaskExecutor now that Streaming uses tasks.

Signed-off-by: AMZN-koppersr <82230785+AMZN-koppersr@users.noreply.github.com>

* Added TaskGraph to the tests that were missing it.

AZ::IO::Streamer now uses TaskGraph so any tests that explicitly sets up AZ::IO::Streamer now first has to setup a TaskExecutor or a TaskGraphSystemComponent.

Signed-off-by: AMZN-koppersr <82230785+AMZN-koppersr@users.noreply.github.com>

* Fixed a race condition within AZ::IO::Streamer.

In rare cases it's possible for a read request to complete before the task thread issuing the request has had a change to update the processing states. In these cases FinalizeReads would not close out the read because it finds the request still waiting to be started. Eventually the task thread does set the processing status, but the scheduler thread won't be woken up. If this was the last read in a chain this can cause a deadlock waiting for the read to complete. This fix sets the state earlier but clears the OVERLAPPED structure so FinalizeReads doesn't complete requests that haven't been issued with the OS yet.

Signed-off-by: AMZN-koppersr <82230785+AMZN-koppersr@users.noreply.github.com>

* Updated Streaming throughput optimizations to compile with TaskGraph updates.

A graphi label is now supplied to each TaskGraph local variable.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>

* Fixed System Component dependencies between Streamer and TaskGraph systems.

The StreamerComponent would create an `AZ:IO::Streamer` instance during its activate function, which constructs an `AZ::IO::StreamerContext`.
The StreamerContext constructor calls `AZ::TaskExecutor::GetInstance()` which returns an lvalue reference to the TaskExecutor.
However the TaskExecutor isn't created unless the TaskGraphSystemComponent has been created.

The StreamerComponent had the "TaskExecutorService" set as a dependent service, but it really is a **required** service.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>

* Fix the error introduced from rebasing.

Signed-off-by: Ross Charles C. <10781734+rossbridger@users.noreply.github.com>

* Fix the bug in LRU algorithm that causes buffer overflow.

Signed-off-by: Ross Charles C. <10781734+rossbridger@users.noreply.github.com>

* Fix benchmark test error.

Signed-off-by: Ross Charles C. <10781734+rossbridger@users.noreply.github.com>

---------

Signed-off-by: AMZN-koppersr <82230785+AMZN-koppersr@users.noreply.github.com>
Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>
Signed-off-by: Ross Charles C. <10781734+rossbridger@users.noreply.github.com>
Co-authored-by: AMZN-koppersr <82230785+AMZN-koppersr@users.noreply.github.com>
Co-authored-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>